backend/requirements.txt
sentence-transformers
flask
flask-cors
pymongo
python-dotenv
beautifulsoup4
langchain
langchain-google-genai
langchain-pinecone
langchain-huggingface
langchain-text-splitters
aiohttp
lxml